Update onnx_api.py
增加 G2PWModel_1.1.zip 文件的判断

+++ GPT_SoVITS/text/g2pw/onnx_api.py
@@ -57,14 +57,18 @@ def download_and_decompress(model_dir: str='G2PWModel/'):
         zip_dir = os.path.join(parent_directory,"G2PWModel_1.1.zip")
         extract_dir = os.path.join(parent_directory,"G2PWModel_1.1")
         extract_dir_new = os.path.join(parent_directory,"G2PWModel")
-        print("Downloading g2pw model...")
-        modelscope_url = "https://paddlespeech.bj.bcebos.com/Parakeet/released_models/g2p/G2PWModel_1.1.zip"
-        with requests.get(modelscope_url, stream=True) as r:
-            r.raise_for_status()
-            with open(zip_dir, 'wb') as f:
-                for chunk in r.iter_content(chunk_size=8192):
-                    if chunk:
-                        f.write(chunk)
+        
+        if os.path.exists(zip_dir):
+            print(zip_dir, "already exists")
+        else:
+            print("Downloading g2pw model...")
+            modelscope_url = "https://paddlespeech.bj.bcebos.com/Parakeet/released_models/g2p/G2PWModel_1.1.zip"
+            with requests.get(modelscope_url, stream=True) as r:
+                r.raise_for_status()
+                with open(zip_dir, 'wb') as f:
+                    for chunk in r.iter_content(chunk_size=8192):
+                        if chunk:
+                            f.write(chunk)
 
         print("Extracting g2pw model...")
         with zipfile.ZipFile(zip_dir, "r") as zip_ref:
